requirements.txt
flask
spotipy
python-dotenv
pandas

